# registry/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API is entrypoint for user requests.\nUser has option to order some products and list created orders.\nAlso, user is able to list his requests to this API.",
        "title": "Registry",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:80",
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "description": "Returns name, version and link to the swagger.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "general"
                ],
                "summary": "Welcome message from microservice.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "auth User ID",
                        "name": "uid",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ServerResponse"
                        }
                    }
                }
            }
        },
        "/order": {
            "post": {
                "description": "Processing user order request. When `uid` param missed - creates new user and wallet as well.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Creates new order.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "auth User ID",
                        "name": "uid",
                        "in": "query"
                    },
                    {
                        "description": "Order",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.OrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Order"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/orders": {
            "get": {
                "description": "Find and return created orders of the user using paging.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Returns list of created orders for all users.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "auth User ID",
                        "name": "uid",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Order"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/orders/{id}": {
            "get": {
                "description": "Find and return created orders of the user using paging.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Returns list of created orders for user.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID to filter orders",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "auth User ID",
                        "name": "uid",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Order"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/requests": {
            "get": {
                "description": "Find and return created requests of the users using paging.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "requests"
                ],
                "summary": "Returns list of created requests for all users.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "auth User ID",
                        "name": "uid",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.UserRequest"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/requests/{id}": {
            "get": {
                "description": "Find and return created requests of the user using paging.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "requests"
                ],
                "summary": "Returns list of created requests for user.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID to filter requests",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "auth User ID",
                        "name": "uid",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.UserRequest"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "x-order": "0"
                }
            }
        },
        "api.ServerResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "x-order": "0"
                },
                "version": {
                    "type": "string",
                    "x-order": "1"
                },
                "swagger": {
                    "type": "string",
                    "x-order": "2"
                }
            }
        },
        "models.Order": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "x-order": "0"
                },
                "user_id": {
                    "type": "string",
                    "x-order": "1"
                },
                "status": {
                    "type": "string",
                    "x-order": "2"
                },
                "amount": {
                    "type": "number",
                    "x-order": "3"
                },
                "timestamp": {
                    "type": "string",
                    "x-order": "4"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.OrderProduct"
                    },
                    "x-order": "5"
                }
            }
        },
        "models.OrderProduct": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "x-order": "0"
                },
                "quantity": {
                    "type": "integer",
                    "x-order": "1"
                }
            }
        },
        "models.UserRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "x-order": "0"
                },
                "user_id": {
                    "type": "string",
                    "x-order": "1"
                },
                "type": {
                    "type": "string",
                    "x-order": "2"
                },
                "path": {
                    "type": "string",
                    "x-order": "3"
                },
                "timestamp": {
                    "type": "string",
                    "x-order": "4"
                }
            }
        },
        "requests.OrderRequest": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.OrderProduct"
                    }
                }
            }
        }
    },
    "x-extension-openapi": {
        "example": "value on a json format"
    }
}